query_viewer/requirements.txt
streamlit>=1.28.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0,<2.3.0
db-dtypes
